requirements.txt
Django==1.5c1
PyYAML==3.10
django-tastypie==0.9.12-alpha
mimeparse==0.1.3
psycopg2==2.4.6
python-dateutil==1.5
wsgiref==0.1.2
